+++ src/pipecat/transports/services/daily.py
@@ -35,6 +35,7 @@ from pipecat.frames.frames import (
     StartFrame,
     TranscriptionFrame,
     TransportMessageFrame,
+    TransportMessageUrgentFrame,
     UserImageRawFrame,
     UserImageRequestFrame,
 )
@@ -64,6 +65,11 @@ VAD_RESET_PERIOD_MS = 2000
 
 @dataclass
 class DailyTransportMessageFrame(TransportMessageFrame):
+    participant_id: str | None = None
+
+
+@dataclass
+class DailyTransportMessageUrgentFrame(TransportMessageUrgentFrame):
     participant_id: str | None = None
 
 
@@ -234,12 +240,12 @@ class DailyTransportClient(EventHandler):
     def set_callbacks(self, callbacks: DailyCallbacks):
         self._callbacks = callbacks
 
-    async def send_message(self, frame: TransportMessageFrame):
-        if not self._client:
+    async def send_message(self, frame: TransportMessageFrame | TransportMessageUrgentFrame):
+        if not self._joined or self._leaving:
             return
 
         participant_id = None
-        if isinstance(frame, DailyTransportMessageFrame):
+        if isinstance(frame, (DailyTransportMessageFrame, DailyTransportMessageUrgentFrame)):
             participant_id = frame.participant_id
 
         future = self._loop.create_future()
@@ -728,21 +734,37 @@ class DailyOutputTransport(BaseOutputTransport):
 
         self._client = client
 
+        # Task to process outgoing messages.
+        self._messages_task = None
+        self._messages_queue = asyncio.Queue()
+
     async def start(self, frame: StartFrame):
         # Parent start.
         await super().start(frame)
         # Join the room.
         await self._client.join()
+        # Start messages task
+        self._messages_task = self.get_event_loop().create_task(self._messages_task_handler())
 
     async def stop(self, frame: EndFrame):
         # Parent stop.
         await super().stop(frame)
+        # Cancel messages task
+        if self._messages_task:
+            self._messages_task.cancel()
+            await self._messages_task
+            self._messages_task = None
         # Leave the room.
         await self._client.leave()
 
     async def cancel(self, frame: CancelFrame):
         # Parent stop.
         await super().cancel(frame)
+        # Cancel messages task
+        if self._messages_task:
+            self._messages_task.cancel()
+            await self._messages_task
+            self._messages_task = None
         # Leave the room.
         await self._client.leave()
 
@@ -750,8 +772,8 @@ class DailyOutputTransport(BaseOutputTransport):
         await super().cleanup()
         await self._client.cleanup()
 
-    async def send_message(self, frame: TransportMessageFrame):
-        await self._client.send_message(frame)
+    async def send_message(self, frame: TransportMessageFrame | TransportMessageUrgentFrame):
+        await self._messages_queue.put(frame)
 
     async def send_metrics(self, frame: MetricsFrame):
         metrics = {}
@@ -776,13 +798,24 @@ class DailyOutputTransport(BaseOutputTransport):
         message = DailyTransportMessageFrame(
             message={"type": "pipecat-metrics", "metrics": metrics}
         )
-        await self._client.send_message(message)
+        await self._messages_queue.put(message)
 
     async def write_raw_audio_frames(self, frames: bytes):
         await self._client.write_raw_audio_frames(frames)
 
     async def write_frame_to_camera(self, frame: OutputImageRawFrame):
         await self._client.write_frame_to_camera(frame)
+
+    async def _messages_task_handler(self):
+        while True:
+            try:
+                message = await self._messages_queue.get()
+                await self._client.send_message(message)
+                self._messages_queue.task_done()
+            except asyncio.CancelledError:
+                break
+            except Exception as e:
+                logger.exception(f"{self} error processing message queue: {e}")
 
 
 class DailyTransport(BaseTransport):